CloudWaterMarking/CloudWaterMarkingPython/requirements.txt
mysql-connector-python
numpy
pycryptodome
jinja2